+++ labelbox/data/serialization/ndjson/converter.py
@@ -6,6 +6,8 @@ from .label import NDLabel
 
 logger = logging.getLogger(__name__)
 
+IGNORE_IF_NONE = ["page", "unit"]
+
 
 class NDJsonConverter:
 
@@ -39,4 +41,8 @@ class NDJsonConverter:
             A generator for accessing the ndjson representation of the data
         """
         for example in NDLabel.from_common(labels):
-            yield example.dict(by_alias=True)
+            res = example.dict(by_alias=True)
+            for k, v in list(res.items()):
+                if k in IGNORE_IF_NONE and v is None:
+                    del res[k]
+            yield res