+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+# Version 3.26.2 (2022-09-06)
+### Added
+* Support for document (pdf) de/serialization from exports
+    * Use the `LBV1Converter.serialize()` and `LBV1Converter.deserialize()` methods
+* Support for document (pdf) de/serialization for imports
+    * Use the `NDJsonConverter.serialize()` and `NDJsonConverter.deserialize()` methods
+
+# Version 3.26.1 (2022-08-23)
+### Changed
+* `ModelRun.get_config()`
+    * Modifies get_config to return un-nested Model Run config
+### Added
+* `ModelRun.update_config()`
+    * Updates model run training metadata
+* `ModelRun.reset_config()`
+    * Resets model run training metadata
+* `ModelRun.get_config()`
+    * Fetches model run training metadata
+    
+### Changed
+* `Model.create_model_run()`
+    * Add training metadata config as a model run creation param
+
 # Version 3.26.0 (2022-08-15)
 ## Added
 * `Batch.delete()` which will delete an existing `Batch`
@@ -663,7 +686,3 @@ a `Label`. Default value is 0.0.
 
 ## Version 2.2 (2019-10-18)
 Changelog not maintained before version 2.2.
-
-### Changed
-* `Model.create_model_run()`
-    * Add training metadata config as a model run creation param